simplify output
